requirements.txt
flask==3.1.0
python-dotenv==1.1.0
pdfplumber==0.11.9
python-docx==1.1.2
pytest==8.3.5
